# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/equipment/":{"get":{"tags":["Оборудование"],"summary":"Get Equipment","description":"Получение списка оборудования","operationId":"get_equipment_equipment__get","parameters":[{"name":"offset","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Offset"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":50,"title":"Limit"}},{"name":"sort_by","in":"query","required":false,"schema":{"$ref":"#/components/schemas/SortEquipment","description":"Поля для сортировки: \n\n            - name: по названию оборудования \n\n            - lab_name: по названию лаборатории \n\n            - status: по статусу \n\n            - start_time: по времени начала бронирования \n\n            - end_time: по времени окончания бронирования \n\n            ","default":"name"},"description":"Поля для сортировки: \n\n            - name: по названию оборудования \n\n            - lab_name: по названию лаборатории \n\n            - status: по статусу \n\n            - start_time: по времени начала бронирования \n\n            - end_time: по времени окончания бронирования \n\n            "},{"name":"order_by","in":"query","required":false,"schema":{"$ref":"#/components/schemas/OrderBy","description":"Сортировать по: \n\n            - asс: Возрастанию \n\n            - desc: Убыванию \n\n            ","default":"asc"},"description":"Сортировать по: \n\n            - asс: Возрастанию \n\n            - desc: Убыванию \n\n            "}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EquipmentObject"},"title":"Response Get Equipment Equipment  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Оборудование"],"summary":"Reserve Equipment","description":"Зарезервировать оборудование","operationId":"reserve_equipment_equipment__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReserveEquipmentRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"EquipmentObject":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"lab_name":{"type":"string","title":"Lab Name"},"student_name":{"type":"string","title":"Student Name"},"status":{"$ref":"#/components/schemas/EquipmentStatus"},"start_time":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Start Time"},"end_time":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"End Time"}},"type":"object","required":["id","name","lab_name","student_name","status","start_time","end_time"],"title":"EquipmentObject"},"EquipmentStatus":{"type":"string","enum":["AVAILABLE","SERVICE","BOOKED"],"title":"EquipmentStatus"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"OrderBy":{"type":"string","enum":["asc","desc"],"title":"OrderBy"},"ReserveEquipmentRequest":{"properties":{"id":{"type":"integer","title":"Id"},"student_name":{"type":"string","title":"Student Name"},"start_time":{"type":"string","format":"date-time","title":"Start Time"},"end_time":{"type":"string","format":"date-time","title":"End Time"}},"type":"object","required":["id","student_name","start_time","end_time"],"title":"ReserveEquipmentRequest"},"SortEquipment":{"type":"string","enum":["name","lab_name","status","start_time","end_time"],"title":"SortEquipment"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}